Add ImageCropper component
